Changed double values to float for optimization, divided constraints in SimulatedObject class

+++ src/Object.cpp
@@ -18,7 +18,7 @@ bool Object::loadObject(const std::string& path)
     std::stringstream ss;
 
     bool first = true;
-    Eigen::Vector3d minVal, maxVal;
+    Eigen::Vector3f minVal, maxVal;
 
     while (std::getline(file, line))
     {
@@ -30,10 +30,10 @@ bool Object::loadObject(const std::string& path)
 
         if (type == "v")
         {
-            double x, y, z;
+            float x, y, z;
             ss >> x >> y >> z;
 
-            auto v = Eigen::Vector3d(x, y, z);
+            auto v = Eigen::Vector3f(x, y, z);
             vertices.emplace_back(v);
 
             if (first)
@@ -59,11 +59,11 @@ bool Object::loadObject(const std::string& path)
 
     // normalize and center
 
-    Eigen::Vector3d center = 0.5 * (minVal + maxVal);
-    Eigen::Vector3d size   = maxVal - minVal;
+    Eigen::Vector3f center = 0.5f * (minVal + maxVal);
+    Eigen::Vector3f size   = maxVal - minVal;
 
-    double maxDim = size.maxCoeff();
-    double scale = 1.0 / maxDim;
+    float maxDim = size.maxCoeff();
+    float scale = 1.0f / maxDim;
 
     for (auto& v : vertices)
     {
